jit: pointer-width exception-class reads, the wasm exception triple, and the exception-edge bridge on every backend
`PyObject` is `{ ob_type, w_class }`, two machine words. Four sites read the
one-word `typeptr` at offset 0 through `*const i64`, so on a 32-bit target the
value carried the adjacent `w_class` in its high half and could never compare
equal to the pending-exception cell, which `jit_exc_raise` publishes at pointer
width. All four now read at pointer width; the reads are byte-identical on
64-bit targets.

The site that reaches generated code is `bridge_subwalk::dispatch_via_miframe`,
whose `exc_edge_class` becomes the bridge-entry GUARD_EXCEPTION constant. On
wasm that guard executed 218889 times and passed 0 times, measured by baking a
`BRIDGE_DIAG` slot address into the emitted guard and storing both compare
operands: runtime `JIT_EXC_TYPE` 0x0000_0000_0083_4658 against a baked
0x00E0_EA40_0083_4658. The three `pyjitpl.rs` sites are the same read; the
third fires once per raising iteration.

The wasm backend no-opped SAVE_EXCEPTION / SAVE_EXC_CLASS / RESTORE_EXCEPTION.
SAVE_EXCEPTION produces a value, and the explicit arm bypassed the
value-producing decline in the `_` fallback, so the local stayed null and the
resumed handler compared against it. Lowered all three against
`x86/assembler.py` genop_save_exc_class, genop_save_exception and
_restore_exception.

With both fixed, `exc_edge_bridge_enabled()` returns `true` for every backend
instead of `cfg!(not(target_arch = "wasm32"))`, and the wasm guard-failure
counts land on the native ones:

  fixture                           wasm before -> after   dynasm
  type_name_surrogate_reject        9464 -> 202            201
  exc_caught_in_callee_return_loop  23748 -> 613           611
  inline_subwalk_property_mutates   23748 -> 613           611
  inline_subwalk_mutating_residual  23950 -> 815           813
  named_reraise_sibling_hot         2649 -> 1712           1712
  sre_pattern_methods               2536 -> 1671           1670
  handler_reraise_second_exc        1285 -> 804            804

`loops_aborted` goes to 0 on all seven; `type_name_surrogate_reject` also drops
`jit_calls` 103624 -> 2248 and `compile_ms` 76 -> 7.8. The seven `.wasm.jitstats`
baselines are re-recorded.

`BRIDGE_DIAG` grows to 30 slots: `cell_set` / `cell_missing` / `cell_rebridge`
separate "the backend accepted a bridge" from "the source guard's dispatch cell
can reach it", which `BRIDGE_OK` alone does not.

The `if !exc_edge_bridge_enabled()` legacy prologue branch in `call_jit.rs` is
now unreachable and is left in place.

+++ majit/majit-metainterp/src/pyjitpl.rs
@@ -9468,7 +9468,12 @@ impl<M: Clone> MetaInterp<M> {
         let exc_class = if result.exception_value.is_null() {
             0
         } else {
-            unsafe { *(result.exception_value.0 as *const i64) }
+            // `typeptr` is one machine word at offset 0, so read it at
+            // pointer width: an i64 read on a 32-bit target would pull the
+            // adjacent header word into the high half and the class value
+            // would never compare equal to a pointer-width one
+            // (`Cpu::cls_of_gcref`, `jit_exc_raise`).
+            unsafe { *(result.exception_value.0 as *const usize) as i64 }
         };
         let exception = ExceptionState {
             exc_class,
@@ -9638,7 +9643,8 @@ impl<M: Clone> MetaInterp<M> {
         let exc_class = if exc_value_ref.is_null() {
             0
         } else {
-            unsafe { *(exc_value_ref.0 as *const i64) }
+            // Pointer-width read — see the `result.exception_value` site.
+            unsafe { *(exc_value_ref.0 as *const usize) as i64 }
         };
         let exception = ExceptionState {
             exc_class,
@@ -9816,7 +9822,8 @@ impl<M: Clone> MetaInterp<M> {
         let exc_class = if exc_value_ref.is_null() {
             0
         } else {
-            unsafe { *(exc_value_ref.0 as *const i64) }
+            // Pointer-width read — see the `result.exception_value` site.
+            unsafe { *(exc_value_ref.0 as *const usize) as i64 }
         };
         let exception = ExceptionState {
             exc_class,
